# api/openapi.yaml
openapi: 3.0.3
info:
  title: 家事 Agent API
  version: 0.1.0
  description: |
    家事 Agent（HomeAgent）后端接口。
    对话为主入口：SSE 流式对话 + 工具调用 + 结果卡片 + 撤销。
    契约先行：改接口先改本文件，再 make generate（CONVENTIONS-backend §1）。
servers:
  - url: /api/v1
    description: 同域反向代理
  - url: http://localhost:8080/api/v1
    description: 本地开发
security:
  - bearerAuth: []
...
paths:
...
  /admin/providers:
    get:
      tags: [admin]
      summary: 供应商配置清单（含脱敏密钥）
      description: |
        api_key 落库加密，此接口只返回脱敏值（如 sk-a***nop）。
        仅供管理端编辑回显，禁止回传明文（ADR 安全约束）。
      operationId: listProviders
      responses:
        '200':
          description: 供应商清单
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Provider'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/providers/{providerId}:
    put:
      tags: [admin]
      summary: 更新供应商配置（加密存密钥）
      description: |
        api_key 传入即加密覆盖；传空串或省略表示不改密钥。
        只有 parent 角色可写（权限双保险：路由层 + 执行层）。
      operationId: updateProvider
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key:
                  type: string
                  format: password
                  description: 明文密钥，服务端加密入库；空串/省略 = 不修改
                base_url:
                  type: string
                  description: 供应商 API 地址，空用供应商默认
                enabled:
                  type: boolean
      responses:
        '200':
          description: 更新后的供应商（脱敏）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/models/{modelId}:
    put:
      tags: [admin]
      summary: 更新模型配置（启停/设默认）
      description: |
        设默认模型是互斥操作：旧默认自动取消。
        只有 parent 角色可写。
      operationId: updateModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                is_default:
                  type: boolean
                  description: 设为全局默认（互斥，旧默认自动取消）
      responses:
        '200':
          description: 更新后的模型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /expenses:
    get:
      tags: [expense]
      summary: 记账流水（按成员隔离）
      operationId: listExpenses
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - in: query
          name: category
          description: 按分类过滤
          schema:
            type: string
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 流水列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expense'
                  next_cursor:
                    type: string
    post:
      tags: [expense]
      summary: 手动记一笔（对话外直接记账，同样幂等可撤销）
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: 已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '409':
          $ref: '#/components/responses/Conflict'
  /expenses/summary:
    get:
      tags: [expense]
      summary: 记账汇总（按分类/按月）
      operationId: expenseSummary
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [month, year]
            default: month
      responses:
        '200':
          description: 汇总
          content:
            application/json:
              schema:
                type: object
                required: [total_cents, categories]
                properties:
                  total_cents:
                    type: integer
                    description: 总支出（分）
                  categories:
                    type: array
                    items:
                      type: object
                      required: [category, total_cents]
                      properties:
                        category:
                          type: string
                        total_cents:
                          type: integer
  /budget:
    get:
      tags: [expense]
      summary: 预算执行情况（query_budget 工具的 HTTP 出口）
      operationId: queryBudget
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: 预算执行
          content:
            application/json:
              schema:
                type: object
                required: [budget_cents, spent_cents]
                properties:
                  budget_cents:
                    type: integer
                  spent_cents:
                    type: integer
                  over_budget:
                    type: boolean
...
  /audit:
    get:
      tags: [admin]
      summary: 审计日志（只增不改，按权限可见）
      operationId: listAudit
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - in: query
          name: tool_name
          schema:
            type: string
        - in: query
          name: permission_denied
          description: 只看越权尝试
          schema:
            type: boolean
      responses:
        '200':
          description: 审计列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  next_cursor:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PageCursor:
      in: query
      name: cursor
      schema:
        type: string
...
    ProviderId:
      in: path
      name: providerId
      required: true
      schema:
        type: string
        format: uuid
    ModelId:
      in: path
      name: modelId
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 幂等冲突（重复提交）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [invalid_input, unauthorized, permission_denied, not_found, conflict, internal]
          description: 程序判断用（apperr.Code）
        message:
          type: string
          description: 面向用户的中文信息（不泄漏堆栈/SQL）
        trace_id:
          type: string
...
    RiskLevel:
      type: string
      enum: [low, medium, high]
    Model:
      type: object
      required: [id, model_name, display_name]
      properties:
        id:
          type: string
          format: uuid
        model_name:
          type: string
          example: gpt-4o
        display_name:
          type: string
          example: GPT-4o
        provider:
          type: string
          example: openai
        is_default:
          type: boolean
    Provider:
      type: object
      required: [id, name, enabled]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: deepseek
          description: 供应商标识（枚举，见 shared ProviderName）
        base_url:
          type: string
          example: https://api.deepseek.com
        api_key_masked:
          type: string
          example: sk-a***nop
          description: 脱敏密钥（明文绝不回传）
        api_key_set:
          type: boolean
          description: 是否已配置密钥（空密钥时 false）
        enabled:
          type: boolean
...
    Expense:
      type: object
      required: [id, member_id, amount_cents, occurred_at]
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        category:
          type: string
          example: 食材
        amount_cents:
          type: integer
          description: 金额（分）
        hint:
          type: string
          example: 买菜
        note:
          type: string
        occurred_at:
          type: string
          format: date-time
        undoable:
          type: boolean
          description: 是否在 24h 撤销窗口内
    ExpenseCreate:
      type: object
      required: [amount_cents, hint]
      properties:
        amount_cents:
          type: integer
          minimum: 1
        hint:
          type: string
        category:
          type: string
        occurred_at:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      required: [id, tool_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        trace_id:
          type: string
        member_id:
          type: string
          format: uuid
        tool_name:
          type: string
        risk:
          $ref: '#/components/schemas/RiskLevel'
        params:
          type: object
          additionalProperties: true
        result:
          type: string
        undone:
          type: boolean
        permission_denied:
          type: boolean
          description: 越权尝试记录
        latency_ms:
          type: integer
        created_at:
          type: string
          format: date-time